+++ src/components/PageLayout/test.js
@@ -1,8 +1,13 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { Button } from '../Button';
-import { Text } from '../Text';
-import { PageLayout } from './index';
+import {
+  Alert,
+  Button,
+  IconSearch,
+  Input,
+  Text,
+  PageLayout
+} from '../../index';
 
 it('PageLayout renders succesfully', () => {
   const tree = renderer.create(
@@ -13,9 +18,32 @@ it('PageLayout renders succesfully', () => {
       />
       <PageLayout
         heading='Cluster'
-        controls={[
+        topRightControls={[
           <Button text='Details' />,
           <Button text='Issues' />
+        ]}
+        wide
+      />
+      <PageLayout
+        heading='Reports list'
+        aboveComponent={({ className }) => (
+          <Alert className={className} type='error'>Hello, I'm error!</Alert>
+        )}
+        topLeftControls={[
+          <Input rightIcon={<IconSearch />} />,
+          <Button text='Info' size='l' />
+        ]}
+        wide
+      />
+      <PageLayout
+        heading='Reports list'
+        topLeftControls={[
+          <Input rightIcon={<IconSearch />} />,
+          <Button text='Info' size='l' />
+        ]}
+        topRightControls={[
+          <Button text='Details' size='l' />,
+          <Button text='Info' size='l' />
         ]}
         wide
       />